Add offset per TT hop (Necessary for correct results in case of multiple GCLs per edge)

+++ src/main/java/dk/smlaursen/TSNCF/architecture/GCL.java
@@ -17,7 +17,7 @@ public class GCL {
 		return aDuration;
 	}
 	
-	public double getFrequency(){
+	public int getFrequency(){
 		return afrequency;
 	}
